# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Tower Go 用户管理系统 API 文档",
        "title": "Tower Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "jeckwell@126.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:10024",
    "basePath": "/api/v1",
    "paths": {
...
        "/purchase-orders": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "采购单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "门店ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "供应商ID",
                        "name": "supplier_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "日期 格式:2024-01-01",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.PurchaseOrder"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "创建采购单",
                "parameters": [
                    {
                        "description": "采购单信息",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreatePurchaseOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "获取采购单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.PurchaseOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "更新采购单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "采购单信息",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdatePurchaseOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "删除采购单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/actions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "获取采购单可用操作",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/by-supplier": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "按供应商分组获取采购单明细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "取消采购单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "取消原因",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "reason": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/complete": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "完成采购单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "采购单管理"
                ],
                "summary": "确认采购单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采购单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.CreatePurchaseOrderItemReq": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "remark": {
                    "type": "string",
                    "maxLength": 200
                }
            }
        },
        "model.CreatePurchaseOrderReq": {
            "type": "object",
            "required": [
                "items",
                "order_date"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.CreatePurchaseOrderItemReq"
                    }
                },
                "order_date": {
                    "description": "格式: 2024-01-01",
                    "type": "string"
                },
                "remark": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "model.PurchaseOrder": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "creator": {
                    "$ref": "#/definitions/model.User"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PurchaseOrderItem"
                    }
                },
                "order_date": {
                    "type": "string"
                },
                "order_no": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "store": {
                    "$ref": "#/definitions/model.Store"
                },
                "store_id": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.PurchaseOrderItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/model.SupplierProduct"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                },
                "supplier": {
                    "$ref": "#/definitions/model.Supplier"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "unit_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "角色代码：admin(总部管理员), store_admin(门店管理员), staff(普通员工)",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "角色描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "status": {
                    "description": "状态：1=启用 0=禁用",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "门店地址",
                    "type": "string"
                },
                "business_hours": {
                    "description": "营业时间",
                    "type": "string"
                },
                "contact_person": {
                    "description": "联系人",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "门店名称",
                    "type": "string"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "状态：1=正常，2=停业",
                    "type": "integer"
                },
                "store_code": {
                    "description": "门店编码 JWXXXX",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Supplier": {
            "type": "object",
            "properties": {
                "contact_email": {
                    "type": "string"
                },
                "contact_person": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "supplier_address": {
                    "type": "string"
                },
                "supplier_code": {
                    "type": "string"
                },
                "supplier_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SupplierCategory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "supplier": {
                    "$ref": "#/definitions/model.Supplier"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SupplierProduct": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.SupplierCategory"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                },
                "spec": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "supplier": {
                    "$ref": "#/definitions/model.Supplier"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdatePurchaseOrderReq": {
            "type": "object",
            "properties": {
                "remark": {
                    "type": "string",
                    "maxLength": 500
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3,
                        4
                    ]
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employee_no": {
                    "description": "工号，6位数字",
                    "type": "string"
                },
                "gender": {
                    "description": "性别：1=男，2=女",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_at": {
                    "description": "记录最后登录时间（仅后端维护）",
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "角色关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Role"
                        }
                    ]
                },
                "role_id": {
                    "description": "--- 权限管理 (RBAC) ---",
                    "type": "integer"
                },
                "status": {
                    "description": "--- 状态与安全管理 ---",
                    "type": "integer"
                },
                "store": {
                    "description": "门店关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Store"
                        }
                    ]
                },
                "store_id": {
                    "description": "--- 门店关联 ---",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}